test run_onlist_region and run_onlist_read
A new accessor function was added to get onlists for the new read
objects in addition to the older by region type.

I also added some type annotations to be more clear that join_onlists
needs a list of Onlist objects to work. (Since we need the full
information to know if we need to download files)