remote onlist download with the kallisto multiple lists onlist format  (#31)
* Support older versions of matplotlib

the spines[["top", "bottom"...]] structure is a relatively recent
update, this allows working with older versions of matplotlib

* Get the test of seqspec check working again.

The refactoring of repositories to split out the example specification
yaml files means we didn't have any local files to try validating.

So I had to use the stub I had added for other tests, however it
needed some updates to be compatible with the library spec version of
the schema.

Also I did some mocking to avoid needing to create test fastq and
barcode files.

* Increase the number of Xs in the random region

The validator now checks that the length of the sequence string is
"X" * max_len characters.

* Update minimal Region tests and add minimal Read tests.

* Make some minimal tests for the seqspec print functions

* update print command to use the replacment assay_id attribute

previously it was assay

* My test assay used custom_primer which didn't have a color.

I randomly picked sea green.

* Implement downloading lists via urls

Also to work with barcode lists hosted by the DACC transparently
decompress gzip files.

The old read_list function took a filename, but I changed it to take
the onlist object so it would have access to the location attribute
to know if it should be reading locally or remotely instead of just
guessing if the filename string started with a scheme url.

* Only return the onlist filename if it a local file

Even if there's one list but it's remote we need to download it and
put into into a local file.

* Add onlist argument to specify combine barcode list file format.

Kallisto has a format where multiple barcode lists are in one file
separated by whitespace. That's different from the more common
cartisean product format where all the lists are crossed with each
other.

This adds the kallisto format as -f multi, and adds an argument for
the current version -f product, but treats it as a default.

* Fix test for project_regions_to_coordinates

* Minimally test RegionCoordinate and  project_regions_to_coordinates

* test run_onlist_region and run_onlist_read

A new accessor function was added to get onlists for the new read
objects in addition to the older by region type.

I also added some type annotations to be more clear that join_onlists
needs a list of Onlist objects to work. (Since we need the full
information to know if we need to download files)