Search for local copies of barcode files first.
This will look in the current directory on the current machine first,
Then it will search the local machine for the file at the full path portion
of the filename, as interpreted by urlparse.
Finally if the location is remote, it will assume that the file is
accessible at that remote URL.

Adding in this local first search is to make pipeline development
easier since it allows other code to handle retrieving the barcode
files for us.