Output filenames instead of read_ids in the format_chromap function
Also as a convenience I added a get_filenames to the Read object.

The original code operated on finding read_ids with the commit that
added passing around the parsed seqspec object I could then take the
read_ids and look for the fastq filenames.

It's untested but in theory if there's multiple File objects you'll
get a comma separated list of the filenames for the command line.